tests/wpt/web-platform-tests/signed-exchange/sxg-hsts.tentative.html

<!DOCTYPE html>
<title>Loading SignedHTTPExchange with strict-transport-security must fail</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  const sxgUrl = get_host_info().HTTPS_ORIGIN +
                 '/signed-exchange/resources/sxg/sxg-hsts.sxg';
  const message = await openSXGInIframeAndWaitForMessage(t, sxgUrl);
  const innerURL = innerURLOrigin() +
                   '/signed-exchange/resources/inner-url.html';
  assert_equals(message.location, innerURL);
  assert_true(message.is_fallback);
}, 'Loading SignedHTTPExchange with an uncached header ' +
   'Strict-Transport-Security must fail and fallback redirect.');

</script>
</body>
